requirements.txt
adversarial-robustness-toolbox==1.6.0
alembic==1.4.2
attrs==20.3.0
autograd==1.3
cachetools==4.1.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
cma==2.7.0
comet-ml==3.8.1
configobj==5.0.6
cssselect==1.1.0
cycler==0.10.0
databricks-cli==0.11.0
decorator==4.4.2
docker==4.2.1
dulwich==0.20.21
entrypoints==0.3
enum34==1.1.10
everett==1.0.3
fets==0.5.3
ffmpeg-python==0.2.0
Flask==1.1.2
Flask-Cors==3.0.8
fsspec==0.7.4
future==0.18.2
gcsfs==0.6.2
gitdb==4.0.5
GitPython==3.1.3
glouton==0.7.0
google-api-core==1.17.0
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gorilla==0.3.0
GPUtil==1.4.0
gunicorn==19.9.0
idna==2.9
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
kaitaistruct==0.8
kiwisolver==1.2.0
llvmlite==0.35.0
lxml==4.5.2
Mako==1.1.3
MarkupSafe==1.1.1
mergedeep==1.3.0
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mlflow==1.8.0
mypy==0.812
mypy-extensions==0.4.3
networkx==2.5.1
numba==0.52.0
numpy>=1.19.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
parsel==1.6.0
patsy==0.5.1
Pillow==8.2.0
polaris-ml==0.9.4
prometheus-client==0.8.0
prometheus-flask-exporter==0.14.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.25.1
pymoo==0.4.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
querystring-parser==1.2.4
reedsolo==0.3
resampy==0.2.2
rsa==4.0
satnogs-decoders==1.7.0
scikit-image==0.18.1
scikit-learn==0.23.2
scipy==1.4.1
selenium==3.141.0
simplejson==3.17.0
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.13
sqlparse==0.3.1
statsmodels==0.12.2
tabulate==0.8.7
threadpoolctl==2.1.0
tifffile==2021.4.8
torchaudio==0.7.2
torchsummary==1.5.1
torchvision==0.8.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.25.9
w3lib==1.22.0
waitress==1.4.4
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
wurlitzer==2.1.0
xgboost==1.1.1
xlrd==1.2.0
torchattacks==2.14.1
autoray
git+https://github.com/RobustBench/robustbench.git@v0.1
thop  # FLOPS computation
pycocotools>=2.0  # COCO mAP
opencv-python>=4.1.2
seaborn>=0.11.0
matplotlib>=3.2.2
